+++ nf_core/pipelines/download/utils.py
@@ -1,20 +1,12 @@
 import contextlib
 import importlib.resources
 import logging
-import os
 import shutil
 import tempfile
 from collections.abc import Generator
-from enum import Enum
 from pathlib import Path
 
 log = logging.getLogger(__name__)
-
-
-class ContainerRegistryUrls(Enum):
-    SEQERA_DOCKER = "community.wave.seqera.io/library"
-    SEQERA_SINGULARITY = "community-cr-prod.seqera.io/docker/registry/v2"
-    GALAXY_SINGULARITY = "depot.galaxyproject.org/singularity"
 
 
 def copy_container_load_scripts(container_system: str, dest_dir: Path, make_exec: bool = True) -> tuple[str, Path]:
@@ -84,21 +76,3 @@ def intermediate_file_no_creation(output_path: Path) -> Generator[Path, None, No
     except:
         tmp.cleanup()
         raise
-
-
-@contextlib.contextmanager
-def intermediate_dir_with_cd(original_dir: Path, base_dir: Path = Path()):
-    """
-    Context manager to provide and change into a tempdir and ensure its removal and return to the
-    original_dir upon exceptions.
-    """
-
-    if not original_dir.is_dir():
-        raise DownloadError(f"Unable to setup temporary dir, original_dir does not exist: {original_dir}")
-
-    tmp = tempfile.TemporaryDirectory(dir=base_dir)
-    try:
-        os.chdir(tmp.name)
-        yield tmp
-    finally:
-        os.chdir(original_dir)