+++ tests/pipelines/lint/test_nextflow_config.py
@@ -34,7 +34,7 @@ class TestLintNextflowConfig(TestLint):
         lint_obj = nf_core.pipelines.lint.PipelineLint(self.new_pipeline)
         lint_obj.load_pipeline_config()
 
-        lint_obj.nf_config["manifest.name"] = "bad_name"
+        lint_obj.nf_config["manifest"]["name"] = "bad_name"
         result = lint_obj.nextflow_config()
         assert len(result["failed"]) > 0
         assert len(result["warned"]) == 0
@@ -45,7 +45,7 @@ class TestLintNextflowConfig(TestLint):
         lint_obj.load_pipeline_config()
 
         lint_obj.release_mode = True
-        lint_obj.nf_config["manifest.version"] = "dev_is_bad_name"
+        lint_obj.nf_config["manifest"]["version"] = "dev_is_bad_name"
         result = lint_obj.nextflow_config()
         assert len(result["failed"]) > 0
         assert len(result["warned"]) == 0
@@ -59,6 +59,7 @@ class TestLintNextflowConfig(TestLint):
             fail_content = re.sub(r"\btest\b", "testfail", content)
         with open(nf_conf_file, "w") as f:
             f.write(fail_content)
+        Path(self.new_pipeline, "conf", "testfail.config").touch()
         lint_obj = nf_core.pipelines.lint.PipelineLint(self.new_pipeline)
         lint_obj.load_pipeline_config()
         result = lint_obj.nextflow_config()
@@ -105,8 +106,12 @@ class TestLintNextflowConfig(TestLint):
         result = lint_obj.nextflow_config()
         assert len(result["failed"]) == 2
         assert (
+            result["failed"][0]
+            == "Config `params.custom_config_base` is not set to `https://raw.githubusercontent.com/nf-core/configs/master`"
+        )
+        assert (
             result["failed"][1]
-            == "Config default value incorrect: `params.custom_config_base` is set as `https://raw.githubusercontent.com/nf-core/configs/master` in `nextflow_schema.json` but is `null` in `nextflow.config`."
+            == "Default value from the Nextflow schema `params.custom_config_base = `https://raw.githubusercontent.com/nf-core/configs/master`` not found in `nextflow.config`."
         )
 
     def test_allow_params_reference_in_main_nf(self):
@@ -155,7 +160,7 @@ class TestLintNextflowConfig(TestLint):
             content = f.read()
             fail_content = re.sub(
                 r"validate_params\s*=\s*true",
-                "params.validate_params = true\ndummy = 0.000000001",
+                "validate_params = true\ndummy = 0.000000001",
                 content,
             )
         with open(nf_conf_file, "w") as f:
@@ -187,7 +192,7 @@ class TestLintNextflowConfig(TestLint):
             content = f.read()
             fail_content = re.sub(
                 r"validate_params\s*=\s*true",
-                "params.validate_params = true\ndummy = 0.000000001",
+                "validate_params = true\ndummy = 0.000000001",
                 content,
             )
         with open(nf_conf_file, "w") as f: